correction des erreurs de LSP

+++ LSP/apres/Shape.java
@@ -1,4 +1,3 @@
-
-public  interface Shape {
-    int getArea();
+public interface Shape {
+    double getArea();
 }
